Most rules implemented

+++ bank-credit/src/main/java/com/sbnz/bankcredit/repository/IAuthorityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sbnz.bankcredit.model.Authority;
 
-public interface IAuthorityRepository extends JpaRepository<Authority, Integer>{
+public interface IAuthorityRepository extends JpaRepository<Authority, Long>{
 
 }
